requirements.txt
absl-py==2.1.0
aenum==3.0.0
affine==2.3.0
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyascii==0.3.2
anyio==4.5.2
appnope==0.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array2gif==1.0.4
asttokens==3.0.0
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==3.2.1
bokeh==3.1.1
boto3==1.17.60
botocore==1.20.60
bqplot==0.12.43
branca==0.8.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==3.1.0
colorful==0.5.4
colour==0.1.5
comm==0.2.2
contourpy==1.1.1
csvkit==2.0.1
cycler==0.10.0
dask==2021.1.0
dataclasses-json==0.5.3
dbfread==2.0.7
debugpy==1.8.9
decorator==4.4.2
defusedxml==0.6.0
distributed==2023.5.0
duckdb==1.1.3
earthengine-api==1.1.5
easydict==1.9
eerepr==0.0.4
entrypoints==0.3
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.0
flatbuffers==24.3.25
folium==0.18.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==0.18.2
gast==0.3.3
geemap==0.34.3
geetools==0.6.7
geocoder==1.38.1
geojson==3.1.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==2.36.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==1.0.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==2.10.0
httpcore==1.0.7
httplib2==0.18.1
httplib2shim==0.0.3
httpx==0.28.0
idna==2.10
imageio==2.9.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.12.3
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.6.3
isodate==0.7.2
jedi==0.18.0
jinja2==3.1.4
jmespath==0.10.0
joblib==1.1.0
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.3
jupyter-console==6.2.0
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.27.3
jupyterlab-widgets==1.0.0
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
leather==0.4.0
libclang==18.1.1
locket==1.0.0
Markdown==3.3.3
MarkupSafe==2.1.5
marshmallow==3.11.1
marshmallow-enum==1.5.1
matplotlib==3.5.0
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.4.3
networkx==2.6.1
nltk==3.2.4
notebook==7.2.2
notebook-shim==0.2.4
numpy==1.24.3
oauthlib==3.1.0
olefile==0.47
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
pandas==1.2.0
pandocfilters==1.4.3
parsedatetime==2.6
parso==0.8.1
partd==1.4.1
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
preprocessing==0.1.13
prettyprinter==0.18.0
prometheus-client==0.9.0
prompt-toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyperclip==1.9.0
pyproj==3.0.0.post1
pyrsistent==0.17.3
pyshp==2.1.3
python-box==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==26.2.0
qtconsole==5.0.1
QtPy==1.9.0
rasterio==1.1.8
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.7
s3transfer==0.4.2
scikit-image==0.18.2
scikit-learn==1.0.2
scipy==1.6.0
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentinelhub==3.2.1
setuptools-scm==8.1.0
Shapely==1.7.1
six==1.15.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.6
sphinx-rtd-theme==0.2.4
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.1
stringcase==1.2.0
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2021.4.8
tinycss2==1.4.0
tomli==2.0.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
traittypes==0.2.1
typing-extensions==4.12.2
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
utm==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xyzservices==2024.9.0
yarl==1.15.2
zict==3.0.0
zipp==3.20.2
